+++ SB03-Findex-Team3/src/main/java/com/sprint/findex/dto/response/cursor/CursorPageResponseAutoSyncConfigDto.java
@@ -1,5 +1,6 @@
-package com.sprint.findex.dto.response;
+package com.sprint.findex.dto.response.cursor;
 
+import com.sprint.findex.dto.response.AutoSyncConfigDto;
 import java.util.List;
 
 public record CursorPageResponseAutoSyncConfigDto(
